// src/hooks/use-pokemons-by-generation/types.ts
export type GenerationVariation = {
  start: number;
  end: number;
};

export type Generations = {
  1: GenerationVariation;
  2: GenerationVariation;
  3: GenerationVariation;
  4: GenerationVariation;
  5: GenerationVariation;
  6: GenerationVariation;
  7: GenerationVariation;
  8: GenerationVariation;
  9: GenerationVariation;
};